+++ src/Pim/Bundle/EnrichBundle/PimEnrichBundle.php
@@ -22,6 +22,7 @@ class PimEnrichBundle extends Bundle
     {
         $container
             ->addCompilerPass(new Compiler\RegisterMassEditOperatorsPass())
-            ->addCompilerPass(new Compiler\RegisterMassEditOperationsPass());
+            ->addCompilerPass(new Compiler\RegisterMassEditOperationsPass())
+            ->addCompilerPass(new Compiler\SerializerPass('pim_internal_api_serializer'));
     }
 }